Goidamerge method & refactoring of LavalandPlanetSystem

+++ Content.Shared/_Lavaland/Procedural/Prototypes/LavalandRuinPoolPrototype.cs
@@ -33,13 +33,13 @@ public sealed partial class LavalandRuinPoolPrototype : IPrototype
     /// Distance in-between ruins.
     /// </summary>
     [DataField]
-    public float RuinDistance = 24;
+    public int RuinDistance = 24;
 
     /// <summary>
     /// Max distance that Ruins can generate.
     /// </summary>
     [DataField]
-    public float MaxDistance = 336;
+    public int MaxDistance = 336;
 
     /// <summary>
     /// List of all grid ruins and their count.
